requirements.txt
python-dotenv==1.0.1
langchain==0.2.2
langchain-community==0.2.3
unstructured==0.14.4
chromadb==0.5.0
transformers==4.39.1
torch==2.2.1
tiktoken==0.7.0
scikit-learn==1.6.1